Populating extra fields on test start
