+++ package.json
@@ -108,7 +108,7 @@
     "@angular/compiler-cli": "^5.2.11",
     "@dvsa/mes-journal-schema": "^1.0.8",
     "@dvsa/mes-search-schema": "^1.0.2",
-    "@dvsa/mes-test-schema": "^1.0.28",
+    "@dvsa/mes-test-schema": "^1.0.29",
     "@hapi/joi": "^15.1.0",
     "@ionic-native-mocks/device": "^2.0.12",
     "@ionic-native-mocks/secure-storage": "^2.0.12",
