+++ src/modules/tests/test-slot-attributes/test-slot-attributes.selector.ts
@@ -11,7 +11,10 @@ export const extractTestSlotAttributes = (slotData): TestSlotAttributes => ({
   slotId: slotData.slotDetail.slotId,
   start: slotData.slotDetail.start,
   specialNeeds: slotData.booking.application.specialNeeds ? true : false,
+  specialNeedsCode: slotData.booking.application.specialNeedsCode,
   vehicleTypeCode: slotData.vehicleTypeCode,
   extendedTest: slotData.booking.application.extendedTest,
   examinerVisiting: slotData.examinerVisiting,
+  previousCancellation: slotData.booking.previousCancellation,
+  entitlementCheck: slotData.booking.application.entitlementCheck,
 });
