--- src/pages/candidate-details/candidate-details.constants.ts
@@ -1,5 +1,0 @@
-export enum SpecialNeedsCode {
-  NONE = 'NONE',
-  YES = 'YES',
-  EXTRA = 'EXTRA',
-}
